% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{a.high}
\alias{a.high}
\alias{a.low}
\alias{b.high}
\alias{b.low}
\title{calculates bounds for modeled parameters}
\usage{
a.high(t_min)

a.low(t_max)

b.high()

b.low(t_max)
}
\arguments{
\item{t_min}{time of first experiemtal time point after inhibition of transcription (not T0)}

\item{t_max}{time of last experimental time point}
}
\value{
returns the lowest/highest parameter values to be used as bounds on modeled parameters
}
\description{
calculates maximum and minimum bounds for parameters alpha and beta based on experimental time points. These are based on the time points in each experiment. For example if an RNA degrades to 0.1% of the initial (T0) amount by 10 min but the 1st time point measured is at 2 h, it will not be possible to estimate that decay rate.
}
\examples{
a.high(7.5)
a.low(480)
b.low(480)
}
