% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.R
\name{aic}
\alias{aic}
\alias{aicc}
\title{Akaike information criterion (with correction)}
\usage{
aic(maxlLik, p)

aicc(maxlLik, p, n)
}
\arguments{
\item{maxlLik}{maximum log likelihood value identified upon model convergence}

\item{p}{number of parameters in the model}

\item{n}{is the total number of observations of a single gene (e.g., 8 time
points X 4 replicates X 4 treatments/genotypes = 128)}
}
\value{
returns the AIC or AICc values
}
\description{
Calculates AIC or AICc.
}
\examples{
aicc(100,5,15)
}
\keyword{internal}
