% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_var.R
\name{fit_var}
\alias{fit_var}
\title{sigma^2 estimation}
\usage{
fit_var(sse, n)
}
\arguments{
\item{sse}{is the minimum SSE (sum of the squared errors) from the slsqp() fitting}

\item{n}{is the total number of observations of a single gene (e.g., 8 time points X 4 replicates X 4 treatments/genotypes = 128)}
}
\value{
returns the sigma squared estimate
}
\description{
Calculates the variance (sigma^2) estimate from the sum of the squared errors from the fit model.
}
\examples{
sig.fun(1,128)
}
\keyword{internal}
