% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_aic.R
\name{fun_aic}
\alias{fun_aic}
\alias{fun_aicc}
\title{Akaike information criterion (with correction)}
\usage{
fun_aic(maxlLik, p)

fun_aicc(maxlLik, p, n)
}
\arguments{
\item{maxlLik}{maximum log likelihood value identified upon model convergence}

\item{p}{number of parameters in the model}

\item{n}{number of measured values for a single gene}
}
\value{
returns the AIC or AICc values
}
\description{
calculates AIC or AICc
}
\examples{
fun_aicc(100,5,15)
}
