% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_nPar.R
\name{fun_nPar}
\alias{fun_nPar}
\title{number of Parameters function}
\usage{
fun_nPar(model, mod, group)
}
\arguments{
\item{model}{model name (e.g. "mod1")}

\item{mod}{two column data.frame with combination of alpha and beta grouping numbers in each model with rownames "modX"}

\item{group}{matrix of all treatment alpha or beta equivalence groups}
}
\value{
returns the integer value of number of parameters in \code{model}
}
\description{
Calculates number of parameters for a specified model given the model parameter constraints.
}
\examples{
fun_nPar("mod1",data.frame("a"=1:5,"b"=rep(2,5),row.names=paste0("mod",1:5)),
         t(matrix(c(1,2,3,4,1,2,2,2,1,1,2,2,1,1,1,2,1,2,1,2,1,2,2,1),nrow=4)))
}
