% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.fun.R
\name{LL.fun}
\alias{LL.fun}
\title{log likelihood}
\usage{
LL.fun(x, y, n)
}
\arguments{
\item{x}{is the minimum SSE from the slsqp() fitting}

\item{y}{is sigma2 cooresponding to the minimum SSE}

\item{n}{is the total number of observations of a single gene (e.g., 8 time
points X 4 replicates X 4 treatments/genotypes = 128)}
}
\value{
returns Log Likelihood
}
\description{
Calculates the log likelihood value from the sum of the squared errors, sigma2,
    and the total number of data points.
}
\examples{
LL.fun(1,1/128,128)
}
