% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.fun.R
\name{sig.fun}
\alias{sig.fun}
\title{sigma^2 estimation}
\usage{
sig.fun(x, n = 128)
}
\arguments{
\item{x}{is the minimum SSE (sum of the squared errors) from the slsqp() fitting}

\item{n}{is the total number of observations 8 time points, 4 replicates, 4 treatments (or genotypes)}
}
\value{
returns the sigma squared estimate
}
\description{
calculates the variance (sigma^2) estimate from the sum of the squared errors from the fit model.
}
\examples{
sig.fun(1,128)
}
