% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.R
\name{cols}
\alias{cols}
\title{Indexes column names with multiple patterns (i.e., multigrep)}
\usage{
cols(df, w, x = NA, y = NA, z = NA)
}
\arguments{
\item{df}{a dataframe with column names to parse}

\item{w, x, y, z}{character string or regular expression passed to grep pattern argument}
}
\value{
returns a vector of integer indices of the column names of \code{df} that match to all of \code{w}, \code{x}, \code{y}, and \code{z}
}
\description{
Identifies dataframe column names that have all of the pattern arguments (up to 4).
}
\details{
Be aware that column data labels that are part of another data label are not advisable (e.g. mut1, mut2, mut1.mut2; cols(df,'mut1') will return indices for both 'mut1' and 'mut1.mut2' labeled columns
}
\examples{
cols(data.frame(xyz=1:5,zay=6:10,ybz=11:15,tuv=16:20),'y','z') ## returns 1 2 3
}
