% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.fun.R
\name{LL.fun}
\alias{LL.fun}
\title{log likelihood}
\usage{
LL.fun(x, y, n = 128)
}
\arguments{
\item{x}{is the minimum SSE from the slsqp() fitting}

\item{y}{is sigma2 cooresponding to the minimum SSE}

\item{n}{is the total number of observations 8 time points, 4 replicates, 4 treatments (or genotypes)}
}
\value{
returns Log Likelihood
}
\description{
calculates the log likelihood value from the sum of the squared errors, sigma2, and the total number of data points
}
\examples{
LL.fun(1,1/128,128)
}
